Move display helpers to `dwave.gate.utils.display`
